cleanup and fixes

+++ VortexEngine/src/Menus/MenuList/ColorSelect.h
@@ -24,7 +24,6 @@ public:
 private:
   enum ColorSelectState
   {
-    // first run after the menu is entered
     STATE_INIT,
     STATE_PICK_SLOT,
     STATE_PICK_HUE1,
@@ -35,7 +34,7 @@ private:
 
   void showSlotSelection();
   void showSelection(ColorSelectState mode);
-  void showFullSet(uint32_t offMs, uint32_t onMs);
+  void showFullSet(uint8_t offMs, uint8_t onMs);
 
   // the options for saturations
   const uint8_t sats[4] = {
